Move to vllm/utils
