+++ benchmarks/kernels/benchmark_moe.py
@@ -9,8 +9,8 @@ import triton
 from ray.experimental.tqdm_ray import tqdm
 from transformers import AutoConfig
 
-from vllm.engine.arg_utils import FlexibleArgumentParser
 from vllm.model_executor.layers.fused_moe.fused_moe import *
+from vllm.utils import FlexibleArgumentParser
 
 
 class BenchmarkConfig(TypedDict):
