+++ benchmarks/overheads/benchmark_hashing.py
@@ -2,7 +2,7 @@ import cProfile
 import pstats
 
 from vllm import LLM, SamplingParams
-from vllm.engine.arg_utils import FlexibleArgumentParser
+from vllm.utils import FlexibleArgumentParser
 
 # A very long prompt, total number of tokens is about 15k.
 LONG_PROMPT = ["You are an expert in large language models, aren't you?"
