+++ tests/async_engine/api_server_async_engine.py
@@ -5,8 +5,9 @@ import uvicorn
 from fastapi.responses import JSONResponse, Response
 
 import vllm.entrypoints.api_server
-from vllm.engine.arg_utils import AsyncEngineArgs, FlexibleArgumentParser
+from vllm.engine.arg_utils import AsyncEngineArgs
 from vllm.engine.async_llm_engine import AsyncLLMEngine
+from vllm.utils import FlexibleArgumentParser
 
 app = vllm.entrypoints.api_server.app
